Replace all occurrences of <%= with <% for erb -x
This is a workaround to deal with the incomplete ERB block expression
that is allowed in Rails but is invalid ERB code. An example of such
code is:

    <%= form_for something do |f| %>
    <% end %>

The first line isn't a complete expression and wouldn't be valid ERB,
but Rails allows it. The previous implementation of the code that dealt
with this non-standard behaviour didn't correctly handle cases where
there were newlines inside the incomplete expression.

This new solution should work better and is similar to other solutions
to the same problem used in other projects.